--- a/benchmark/python/ffi/benchmark_ffi.py
+++ b/benchmark/python/ffi/benchmark_ffi.py
@@ -83,6 +83,10 @@ def prepare_workloads():
     OpArgMngr.add_workload("may_share_memory", pool['2x3'][:0], pool['2x3'][:1])
     OpArgMngr.add_workload("roll", pool["2x2"], 1, axis=0)
     OpArgMngr.add_workload("rot90", pool["2x2"], 2)
+    OpArgMngr.add_workload("array_split", pool['2X2'], 2, axis=1)
+    OpArgMngr.add_workload("vsplit", pool['2X2'], 2)
+    OpArgMngr.add_workload("hsplit", pool['2X2'], 2)
+    OpArgMngr.add_workload("dsplit", pool['2X2x2'], 2)
 
 
 def benchmark_helper(f, *args, **kwargs):